refactor: reorder dto packages and delete unneeded data class.
