+++ identity/src/main/kotlin/net/thechance/identity/api/dto/auth/AuthResponse.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.dto
+package net.thechance.identity.api.dto.auth
 
 data class AuthResponse(
     val accessToken: String,
