+++ identity/src/main/kotlin/net/thechance/identity/api/dto/auth/RefreshTokenRequest.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.dto
+package net.thechance.identity.api.dto.auth
 
 import jakarta.validation.constraints.NotBlank
