+++ identity/src/main/kotlin/net/thechance/identity/api/dto/otp/RequestOtpResponse.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.dto
+package net.thechance.identity.api.dto.otp
 
 data class RequestOtpResponse(
     val sessionId: String,
