+++ identity/src/main/kotlin/net/thechance/identity/api/dto/otp/VerifyOtpRequest.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.dto
+package net.thechance.identity.api.dto.otp
 
 import jakarta.validation.constraints.NotBlank
 import org.hibernate.validator.constraints.Length
@@ -7,6 +7,7 @@ import org.hibernate.validator.constraints.UUID
 data class VerifyOtpRequest(
     @field:Length(min = 6, max = 6, message = "otp must be 6 digits")
     val otp: String,
+
     @field:NotBlank(message = "sessionId must not be blank")
     @field:UUID(message = "sessionId must be in a valid UUID format")
     val sessionId: String,
