+++ identity/src/main/kotlin/net/thechance/identity/api/dto/password/ChangePasswordRequest.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.dto
+package net.thechance.identity.api.dto.password
 
 import org.hibernate.validator.constraints.Length
 
@@ -12,5 +12,3 @@ data class ChangePasswordRequest(
     @field:Length(min = 8, message = "confirm password must be more than or equals 8 characters")
     val confirmPassword: String,
 )
-
-data class ChangePasswordResponse(val message: String)
