+++ identity/src/main/kotlin/net/thechance/identity/api/dto/password/ChangePasswordResponse.kt
@@ -1,0 +1,3 @@
+package net.thechance.identity.api.dto.password
+
+data class ChangePasswordResponse(val message: String)
