+++ identity/src/main/kotlin/net/thechance/identity/api/dto/password/ResetPasswordRequest.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.dto
+package net.thechance.identity.api.dto.password
 
 import jakarta.validation.constraints.NotBlank
 import org.hibernate.validator.constraints.Length
