+++ identity/src/main/kotlin/net/thechance/identity/api/dto/profile/ProfileResponse.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.dto
+package net.thechance.identity.api.dto.profile
 
 data class ProfileResponse(
     val id: String,
