+++ identity/src/main/kotlin/net/thechance/identity/api/mapper/AddressMapper.kt
@@ -1,8 +1,8 @@
 package net.thechance.identity.api.mapper
 
-import net.thechance.identity.api.dto.AddressResponse
-import net.thechance.identity.api.dto.CreateAddressRequest
-import net.thechance.identity.api.dto.UpdateAddressRequest
+import net.thechance.identity.api.dto.address.AddressResponse
+import net.thechance.identity.api.dto.address.CreateAddressRequest
+import net.thechance.identity.api.dto.address.UpdateAddressRequest
 import net.thechance.identity.entity.Address
 import net.thechance.identity.service.model.Address as AddressModel
 
@@ -19,20 +19,20 @@ fun Address.toResponse(): AddressResponse {
 
 fun CreateAddressRequest.toAddressModel(): AddressModel {
     return AddressModel(
-        latitude = this.latitude,
-        longitude = this.longitude,
-        addressLine = this.addressLine,
-        addressType = this.addressType,
+        latitude = latitude,
+        longitude = longitude,
+        addressLine = addressLine,
+        addressType = addressType,
         isActive = false
     )
 }
 
 fun UpdateAddressRequest.toAddressModel(): AddressModel {
     return AddressModel(
-        latitude = this.latitude,
-        longitude = this.longitude,
-        addressLine = this.addressLine,
-        addressType = this.addressType,
-        isActive = this.isActive
+        latitude = latitude,
+        longitude = longitude,
+        addressLine = addressLine,
+        addressType = addressType,
+        isActive = isActive
     )
 }
