+++ identity/src/main/kotlin/net/thechance/identity/service/AuthenticationService.kt
@@ -1,6 +1,6 @@
 package net.thechance.identity.service
 
-import net.thechance.identity.api.dto.AuthResponse
+import net.thechance.identity.api.dto.auth.AuthResponse
 import net.thechance.identity.entity.User
 import net.thechance.identity.entity.LoginLog
 import net.thechance.identity.exception.InvalidCredentialsException
